fix: changelog release entries (#62)

+++ scripts/release.mts
@@ -8,26 +8,18 @@ import _readChangesets from '@changesets/read'
 const readChangesets =
   _readChangesets.default as typeof import('@changesets/read/dist/declarations/src/index.js')['default']
 
+// These paths are relative to the git root
+const ASSETS_DIR = './assets'
+const CHANGELOG_PATH = './CHANGELOG.md'
+
 async function main() {
-  if (!process.env.DEBUG) {
-    $.verbose = false
-  }
-  await checkExternalScriptDependencies()
-  await runAtGitRoot()
-  // These values are relative to the git root
-  const packDir = './assets'
-  const changelogPath = './CHANGELOG.md'
+  await setup()
   await maybeConsumeVersions()
-
-  const { gitTag, filename } = await pkgToTarball(packDir)
-
-  if (await checkIfReleaseExists(gitTag)) {
-    warn(`Release under tag "${gitTag}" already exists, skipping...`)
-    return
-  }
-
-  await createGithubRelease(packDir, filename, gitTag, changelogPath)
+  const { gitTag, filename } = await pkgToTarball(ASSETS_DIR)
+  const releaseChangelog = await createReleaseChangelog(CHANGELOG_PATH, gitTag)
+  await createGithubRelease(ASSETS_DIR, filename, gitTag, releaseChangelog)
 }
+
 main()
 
 interface NpmPack {
@@ -42,25 +34,91 @@ interface NpmPack {
   files: { path: string; size: number; mode: number }[]
 }
 
+/**
+ * Creates a Github release through the Github CLI, if the release doesn't already exist.
+ * @param packDir The directory where the tarball is located
+ * @param filename The filename of the tarball
+ * @param gitTag The git tag to create the release under
+ * @param changelog The contents of the changelog to include in the release notes
+ */
 async function createGithubRelease(
   packDir: string,
   filename: string,
   gitTag: string,
-  changelogPath: string,
+  changelog: string,
 ) {
+  if (await checkIfReleaseExists(gitTag)) {
+    warn(`Release under tag "${gitTag}" already exists, skipping...`)
+    return
+  }
+
   const asset = `${packDir}/${filename}`
-  log(
-    `Creating release with tag "${gitTag}" with asset "${asset}" and changelog of ${changelogPath}`,
-  )
-  await $`gh release create ${gitTag} ${packDir}/${filename} -F ${changelogPath}`
+  log(`Creating release with tag "${gitTag}" with asset "${asset}."`)
+  await $`echo ${changelog} | gh release create ${gitTag} ${asset} --notes-file -`
+
   log(`Cleaning up...`)
   log(`Removing ${packDir}`)
   fs.remove(packDir)
 }
 
+/**
+ * Creates a modified CHANGELOG.md which excludes older version notes if necessary. Otherwise,
+ * will return the original changelog contents if no modification was necessary.
+ * This code will no-op and return the original changelog path if:
+ * - No versions are found in the changelog
+ * - No other versions are found in the changelog
+ * - Changelog doesn't match expected format
+ * @param changelogPath The path to the changelog
+ * @param gitTag The git tag to include changes for
+ */
+async function createReleaseChangelog(changelogPath: string, gitTag: string) {
+  log(`Modifying changelog to only include changes for ${gitTag}...`)
+  const changelog = await fs.readFile(changelogPath, 'utf-8')
+
+  // git tag version to changelog version header format (v0.0.0-0 -> ## 0.0.0-0)
+  const version = `## ${gitTag.substring(1)}`
+  // Pattern for "## x.y.z" with an optional snapshot release suffix (-abc123)
+  const changelogHeaderVersionPattern =
+    /^## [0-9]{1,3}.[0-9]{1,3}.[0-9]{1,3}(?:-[0-9a-z]{1,15}){0,1}$/gm
+
+  const matchedVersions: { match: string; index: number }[] = [
+    ...changelog.matchAll(changelogHeaderVersionPattern),
+  ]
+    .map((m) => {
+      return { match: m[0], index: m.index }
+    })
+    .filter((m): m is { match: string; index: number } => m.index !== undefined) // filter out undefined indexes
+    .sort((a, b) => a.index! - b.index!) // sort ascending by index (top to bottom)
+
+  if (matchedVersions.length === 0) {
+    warn(`No versions found in changelog, skipping changelog modification.`)
+    return changelog
+  } else if (matchedVersions[0].match !== `${version}`) {
+    log(
+      `First changelog version entry is ${matchedVersions[0].match} and not ${version}, skipping changelog modification.`,
+    )
+    return changelog
+  } else {
+    // First index is now guaranteed to be the one matching the git tag
+    // We want to trim everything after (and including) the next version header
+    const trimIndex = matchedVersions[1]?.index ?? changelog.length
+    return changelog.substring(0, trimIndex)
+  }
+}
+
+/**
+ * If there are changesets (changeset files in .changeset directory), consume them and create a snapshot release.
+ * That means, there's no changeset files at HEAD. This happens when:
+ * 1. The changesets PR was merged (bumping versions and removing all changesets)
+ *     - Create a full release
+ * 2. No changesets existed, and another commit was added which didn't include a changeset.
+ *     -  No-op because a release/tag should already exist ass they're derived from the version in each package.json
+ */
 async function maybeConsumeVersions() {
-  const hasChanges = await readChangesets('.').then((sets) => sets.length > 0)
-  if (!hasChanges) {
+  const hasChangesets = await readChangesets('.').then(
+    (sets) => sets.length > 0,
+  )
+  if (!hasChangesets) {
     log(`Attempting to create release....`)
     return
   }
@@ -69,22 +127,9 @@ async function maybeConsumeVersions() {
   await $`yarn changeset version --snapshot`
 }
 
-async function runAtGitRoot() {
-  const gitRoot = await $`git rev-parse --show-toplevel`
-  cd(gitRoot.stdout.trimEnd())
-}
-
-async function checkExternalScriptDependencies() {
-  try {
-    await which('gh')
-  } catch {
-    error(
-      `This script requires "gh" to be installed, see: https://github.com/cli/cli/releases`,
-    )
-    process.exit(1)
-  }
-}
-
+/**
+ * Checks if a release already exists under the given git tag.
+ */
 async function checkIfReleaseExists(gitTag: string): Promise<boolean> {
   return await $`gh release view ${gitTag}`
     .then((p) => {
@@ -111,12 +156,42 @@ async function pkgToTarball(packDir: string) {
   return { gitTag, filename }
 }
 
+async function setup() {
+  // Set zx verbosity based on DEBUG env var
+  if (!process.env.DEBUG) {
+    $.verbose = false
+  }
+  await runAtGitRoot()
+  await checkExternalScriptDependencies()
+}
+
 /**
- * Helper functions
+ * Change the working directory to the git root.
  */
+async function runAtGitRoot() {
+  const gitRoot = await $`git rev-parse --show-toplevel`
+  cd(gitRoot.stdout.trimEnd())
+}
+
+/**
+ * Check that external dependencies are installed.
+ * Currently only checks for "gh" (Github CLI).
+ */
+async function checkExternalScriptDependencies() {
+  try {
+    await which('gh')
+  } catch {
+    error(
+      `This script requires "gh" to be installed, see: https://github.com/cli/cli/releases`,
+    )
+    process.exit(1)
+  }
+}
+
 function log(...params: Parameters<typeof chalk['gray']>): void {
   console.log(chalk.gray(...params))
 }
+
 function debug(...params: Parameters<typeof chalk['blue']>): void {
   if (process.env.DEBUG) {
     console.debug(chalk.blue('DEBUG:'), chalk.blue(...params))
